examples/requirements.txt
aiohttp==3.7.4.post0
requests==2.26.0
python-dotenv>=0.18.0
ucube==0.0.1
